pyproject.toml
[project]
name = "image-processing-on-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "requests>=2.32.4",
]
